Set up automatic deployment of santiment projects
Normally we need to specify (using a hash string) the exact version of
each package that we are using in our nix files. This is good for
third-party packages, but it slows down development for our own
packages.

Here we have implemented a way to directly use the latest commit of a
specified branch when building santiment packages from santiment
repos without specifying the hash of the commit. The branch which is
tracked is taken from the ./deployment_environment file. This file
should contain the name of the current nixpkgs branch. So the "master"
branch for nixpkgs will build the "master" branches for other
santiment repos, the "stage" branch will build "stage" branches and
the "production" branch will build "production" branches of packages.

In this way we can get the master/stage/production Santiment
environment by pointing to the master/stage/produciton branch of
santiment/nixpkgs. Deploying from one of those branches will always
pull the latest commit from the respective branches of the other
santiment repos.

The idea was taken:
from http://chriswarbo.net/projects/nixos/useful_hacks.html